♻️(frontend) refactor room factory
In order to have access to the factory of the rooms in the app, we must modify
the factory and the imports
