+++ src/frontend/magnify/packages/components/src/components/rooms/RoomConfig/RoomConfig.test.tsx
@@ -1,11 +1,12 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import React from 'react';
+import { createRandomRoom } from '../../../factories/rooms';
 import { MagnifyTestingProvider } from '../../app';
 import { RoomConfig } from './RoomConfig';
 
 describe('RoomConfig', () => {
   it('should fetch data from the controller', async () => {
-    render(<RoomConfig room={undefined} />, { wrapper: MagnifyTestingProvider });
+    render(<RoomConfig room={createRandomRoom()} />, { wrapper: MagnifyTestingProvider });
 
     // Wait for loading to finish
     const checkboxes = screen.getAllByRole('checkbox');
